Mixins cleanup [skip netlify]
The mixins file contain many "one lines" mixins
that don't help and serve as bloat. So instead of
using includes to invoke the basic mixins, is
better to use a straightforward method.

Fixed the wrong values in the duration parameter
on the animation and transition mixins and added
a correct variable to be used.

The absoute mixin was removed because it was a
leftover from the old design.