Fixed netlify.toml errors [skip netlify]
Replaced build command because it was outdated and returned a error when
tried to build.

The headers "Content-Security-Policy (CSP)",
"Cross-Origin-Opener-Policy (COOP)",
"Cross-Origin-Embedder-Policy (COEP)",
"Cross-Origin-Resource-Policy (CORP)" and "X-XSS-Protection" are moved
from all files to only in .html since it was unnecessary to different
file formats.

Removed deprecated rule "Expect-CT".

Removed "x-frame-options" as the CSP option frame-ancestors can achieve
the same with consistent support and stronger checks.

Fixed Permissions-Policy as the argument used was wrong and removed
outdated options.

Prettyfied Report-To rule for better visualization.

Changed for rule from site.webmanifest to *.webmanifest to apply to all
.webmanifest files.

Changed cache duration of images, fonts and manifests.

Added Content-Type and charset to manifests and css files.

Added immutable option to cache-control of images, fonts and manifests.

Forced a redirect when http status return 404.

Removed unused plugins and contexts environments.