Fix loading in banner images and wrong data
By now the post banner images has been using the loading="lazy" method,
but it should be using the loading="eager" because the image is on the
viewport and must load faster. Also rewrite some variables in the image
code to be more readable.

Replaced the image in site.json with the new thumbnail file, also
removed the behance link because it's not active.