Fix scss undeclared variables and minor typo [skip netlify]
The modified scss files was not updated to add the prefix of the
imported variable or the mixin used and added a declaration of the
module imported at the beginning of the file.

Fix minor typo in typography.scss with trailing space at path and fix
missign closing bracket in contact.liquid.